requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
blas=1.0=mkl
bottleneck=1.3.5=py39h7deecbd_0
brotlipy=0.7.0=py39h27cfd23_1003
ca-certificates=2022.10.11=h06a4308_0
certifi=2022.9.24=py39h06a4308_0
cffi=1.15.1=py39h74dc2b5_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py39h06a4308_0
cryptography=38.0.1=py39h9ce1e76_0
dataclasses=0.8=pyh6d0b6a4_7
fftw=3.3.9=h27cfd23_1
flask=2.1.3=py39h06a4308_0
flask_cors=3.0.10=pyhd3eb1b0_0
future=0.18.2=py39h06a4308_1
idna=3.4=py39h06a4308_0
importlib-metadata=4.11.3=py39h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
itsdangerous=2.0.1=pyhd3eb1b0_0
jinja2=3.1.2=py39h06a4308_0
joblib=1.1.1=py39h06a4308_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
markupsafe=2.1.1=py39h7f8727e_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
ncurses=6.3=h5eee18b_3
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
numexpr=2.8.3=py39h807cd23_0
numpy=1.23.3=py39h14f4228_1
numpy-base=1.23.3=py39h31eccc5_1
openssl=1.1.1s=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.4=py39h6a678d5_0
pip=22.2.2=py39h06a4308_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py39h06a4308_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.13=haa1d7c7_2
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.10.2=cpu_py39hfa7516b_0
pytz=2022.1=py39h06a4308_0
readline=8.2=h5eee18b_0
requests=2.28.1=py39h06a4308_0
scikit-learn=1.1.3=py39h6a678d5_0
scipy=1.9.3=py39h14f4228_0
setuptools=65.5.0=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.39.3=h5082296_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
typing-extensions=4.3.0=py39h06a4308_0
typing_extensions=4.3.0=py39h06a4308_0
tzdata=2022f=h04d1e81_0
urllib3=1.26.12=py39h06a4308_0
werkzeug=2.0.3=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.6=h5eee18b_0
zipp=3.8.0=py39h06a4308_0
zlib=1.2.13=h5eee18b_0
